Renamed class SkipPolicyVarification to InvalidVariantSkipPolicy. Modified tests.
